+++ phpmyfaq/translations/language_de.php
@@ -1246,12 +1246,6 @@ $PMF_LANG['rightsLanguage::delete_section'] = 'Bereiche löschen';
 $PMF_LANG['rightsLanguage::administrate_sections'] = 'Bereiche administrieren';
 $PMF_LANG['rightsLanguage::administrate_groups'] = 'Gruppen administrieren';
 $PMF_LANG['ad_group_rights'] = 'Rechte der Gruppe';
-$PMF_LANG['ad_menu_meta'] = 'Template-Metadaten';
-$PMF_LANG['ad_meta_add'] = 'Template-Metadaten hinzufügen';
-$PMF_LANG['ad_meta_page_id'] = 'Seitentyp';
-$PMF_LANG['ad_meta_type'] = 'Inhaltstyp';
-$PMF_LANG['ad_meta_content'] = 'Inhalt';
-$PMF_LANG['ad_meta_copy_snippet'] = 'Code-Snippet für Templates kopieren';
 $PMF_LANG['rightsLanguage::viewadminlink'] = 'Link zur Administration sichtbar';
 
 // added v3.0.0-beta.3 - 2019-09-22 by Thorsten